Convert scripts to also run with dash
I am converting the shell-scripted tap tests to also run using the dash
and ksh shell, fixing some other problems on the way as well.

Two of the tests' data files contained '\\1' in an sed command. This
created char(1) with the dash, '\1' with the bash. I found it easier to
replace the sed command with a different one than figuring out how to
solve this escape problem so it works for both shells.