nwfilter: Adapt test program and cases to recent iptables
Recent iptables fixes a lot of issues with missing spaces and
other information that was previously not reported properly. To make
the test program and test cases work on old and newer installations
of iptables tools, some adjustments need to be made.

Fix a 'file not found error' when running this tool from the shell
directly.