nwfilter: probe for inverted ctdir
Linux netfilter at some point inverted the meaning of the '--ctdir reply'
and newer netfilter implementations now expect '--ctdir original'
instead and vice-versa.
We probe for this netfilter change by checking the version of the
running kernel to see whether libvirt is also using the inverted logic
of the --ctdir parameter (for Linux kernel version >= 2.6.39).

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>