hooks/051-daemon-hook: Return 0 after connection reset
Despite running all subtests successfully, hooks/051-daemon-hook.t ends
with a return code of 141. This is due to the connection to tck not being
cleaned up properly after libvirtd is restarted in the middle of the test.

Ignoring the SIGPIPE and specifically undefining $tck allows the test to
complete with a return code of 0 (which is helpful when running the full
libvirt-tck kit).