create vms that use virt-builder images with no graphics and virtio-net
This is consistent with what is used to create the disk image supplied
by virt-builder. It doesn't currently affect the outcome of the test,
but it's possible that in the future it could.

In particular, the network device name changes depending on whether or
not there is a graphics device (due to change in PCI address). The
virt-builder image is create by installing Fedora on a guest that has
no graphics card, so it sees the network device as "ens2" and creates
an appropriate ifcfg-ens2, but if you boot the image with a graphics
card, then the network device will be named "ens3". It turns out that
NetworkManager is enabled by default on the virt-builder images, and
NetworkManager will listen for dhcp on *all* interfaces (not just
those with a configuration file), so networking still functions on the
guest, but if a test script were to try to use "ifdown ens3" (for
example) that wouldn't work.

The best solution may be to have the image set "biosdevnames=0
net.ifnames=0" on the kernel commandline (so that the netdev is always
called "eth0" regardless of its PCI address), but that would also need
to be done in the kickstart file used to create the image, which is
done offline by libguestfs people, so it's not something we have
control over here.

Signed-off-by: Laine Stump <laine@laine.org>
Reviewed-by: Daniel P. Berrange <berrange@redhat.com>