Fix no-arp-spoof test script generation
I'm not sure if or how this ever worked before - what is supposed to
happen is that the shell script in $cmdfile is put into the file
/test.sh on the guest, then /test.sh is executed. But the code was
instead trying to directly execute the text of the entire script as a
single command and write the generated output to /test.sh, then
execute that.

putting echo " ..... " around the whole thing fixed it.

Signed-off-by: Laine Stump <laine@laine.org>
Reviewed-by: Daniel P. Berrange <berrange@redhat.com>