default to no nwfilter in test domains
When the nwfilter tests were added, the config for *all* test domains
was changed to apply the "clean-traffic" filter. This doesn't have an
ill effect on any test that I know of, but if nwfilter code happened
to be broken in just the right way, other tests that might be
unrelated could fail as well.

This patch changes the default nwfilter from "clean-traffic" to
nothing, and explicitly sets clean-traffic for those tests that
require it.

Signed-off-by: Laine Stump <laine@laine.org>
Reviewed-by: Stefan Berger <stefanb@linux.vnet.ibm.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>