Fix ebtables check in no-mac-broadcast test
Once the correct filter is enabled for the no-mac-broadcast test, the
original test to validate ebtables rules is no longer correct - it was
checking for the presence of the test guest's MAC address in the
ebtables output on the host, but the no-mac-broadcast filter doesn't
have the guest's MAC address anywhere. This patch changes the code to
look for "-d Broadcast -j DROP", which actually is added to ebtables
for no-mac-broadcast.

Signed-off-by: Laine Stump <laine@laine.org>
Reviewed-by: Daniel P. Berrange <berrange@redhat.com>