filter for proper IP address in tcpdump looking for broadcast packet
In the no-mac-broadcast test, a ping is sent to 192.168.122.255, but
tcpdump is set to look for packets with a destination IP of
255.255.255.255. Change it to check for the correct IP address and
also for mac broadcast (which is what the no-mac-broadcast filter
actually looks at). This should eliminate the "false success" that was
happening because tcpdump wasn't actually seeing the broadcast packet
the guest was sending, as well as catching the "false failure" caused
by tcpdump seeing other traffic from the guest unrelated to the test
(which happened to be broadcasts sent to 255.255.255.255).

Signed-off-by: Laine Stump <laine@laine.org>
Review-by: Stefan Berger <stefanb@linux.vnet.ibm.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>