nwfilter tests: remove all hardcoded references to 192.168.122 network
The nwfilter tests have a few places that hardcode 192.168.122 as the
address of libvirt's default network. Remove all of these and replace
them with addresses that are dynamically determined based on
get_network_ip().

(This will have the immediate effect of helping the tests to succeed
when libvirt-tck is run in a virtual machine, since virtual machines
often have their default network set to a different subnet (in order
to avoid conflict with the L0 host's default network)).

Signed-off-by: Laine Stump <laine@laine.org>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>