require libguestfs-tool package
The nwfilter tests use virt-builder to download and configure a guest
appliance, but we don't require the libguestfs-tools package. Up until
now it's just happened that it was always already installed on the
test systems, so it wasn't a problem, but it's annoying to have to
type the one extra install command when setting up a new machine to
run the tests.

Signed-off-by: Laine Stump <laine@laine.org>
Reviewed-by: Erik Skultety <eskultet@redhat.com>