Don't assume the domain's network interface tap device is named "vnet0"
Some of the tests perform operations on the tap device of the test
domain's network interface. They mostly assume that this tap device is
named "vnet0", which is the case if there is no other domain running
on the host, but isn't true if some other domain was running.

This patch adds a utility function
NetworkHelpers::get_first_interface_target_dev($dom), which learns the
name of the tap device for the first interface in the domain, and uses
the result of that function instead of "vnet0".

Signed-off-by: Laine Stump <laine@laine.org>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>