scripts: shorten netdev/domain name to prevent overflow
When new pids are more than 9 digits long

The nwfilter tests use tck-test${pid} as the auto-generated name for
their test domains (and therefore the tap devices created for those
test domains). When the pid is > 9 digits long, the total length of
the name is > 16, which is the maximum length for any network device
name (IFNAMSIZ), and this causes failure to create the tap device,
thus failure of the test:

  /home/laine/devel/libvirt-tck/scripts/nwfilter/050-apply-verify-host.t ..
  error: Failed to start domain tck-test10513851
  error: Network interface name 'tck-test10513851' is too long:
         Numerical result out of range

The solution is to just remove "test" from the name template, so pid
can be up to 12 digits long before the test fails.

Signed-off-by: Laine Stump <laine@laine.org
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>