network tests: change network address to avoid commonly-used subnet
My own machine has a virtual network on 192.168.123.0/24, and I've
been told that it's a common selection for people who create a second
network after libvirt's default (and in the past was used as a manual
alternative for the default network itself when 192.168.122.0/24
casued a conflict). Changing to 192.168.97.0/24 also isn't foolproof,
but it's much less likely to encounter a conflict leading to a failed
test.

Signed-off-by: Laine Stump <laine@laine.org
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>