eliminate use of deprecated brctl in network tests
brctl is part of the bridge-utils package, which has been deprecated /
replaced by the "ip" and "bridge" commands in the iproute package in
all modern distros. This patch removes the few usages of brctl in
libvirt-tck's network tests, replacing it with a functional equivalent
that uses "ip link show type bridge" instead.

Signed-off-by: Laine Stump <laine@laine.org>
Reviewed-by: John Ferlan <jferlan@redhat.com>