fix selinux/100-static-relabel-no test
This test disables auto-relabeling so that it can test if setting a
user-defined label in the config will work. The problem is that the
disk image used to boot is then not labelled correctly, and qemu can't
open it.

To circumvent this failure (which is *not* of the functionality we're
testing), this patch just temporarily sets SELinux to permissive mode
during the test, then sets it back during cleanup.

Signed-off-by: Laine Stump <laine@Laine.org>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>