hooks: Add systemd detection
The hooks assume System V. On RPM-based distros, there's the initscripts
package introducing the 'service' command mapping the old style syntax
to the systemd format. However, we can't assume this will be the case
all the time, so some kind of detection of the init system would prevent
test failures originating in the hooks because of the 'service' command
missing.

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>