scripts: fix disk media change test
Since the introduction of the 'index' attribute of a disk's
<source> element, test 207-disk-media-change.t is failing. The
test is a bit flawed in that it compares initial and final domXML
after changing the media of a cdrom disk device. The test only
needs to check the final domXML to ensure the correct <source>
path exists at the end of the test.

Signed-off-by: Jim Fehlig <jfehlig@suse.com>
Reviewed-by: Erik Skultety <eskultet@redhat.com>
Reviewed-by: Daniel P. Berrangé <berrange@redhat.com>