nwfilter: ipv6-test: Adjust the ebtables regex output matching
So, in my commit 8d0c9600, I adjusted the expected ebtables output
against which we diff the actuall ebtables output by explicitly adding a
/128 IPv6 prefix to src/dst addresses that were supposed to have that.
The problem is that this was fixed against Fedora-31 which changed the
ebtables output for some reason. That version of ebtables came from the
"ebtables-legacy" whereas on Fedora-32 and CentOS-8  ebtables points to
a busybox binary instead and the output matches the one prior to
Fedora-31.
Now, the correct fix would be a complete overhaul and finding a way not
to rely on stdout of a program in the test suite which is very fragile,
but that would require re-working the whole nwfilter test suite. Let's
provide a quick fix in the meantime by adjusting the regex so that the
/128 prefix is always added if the ebtables output doesn't use it
explicitly.
Note that we don't have to bother with whitespaces in the regexes, since
we're diffing with whitespaces stripped (diff -w).

Signed-off-by: Erik Skultety <eskultet@redhat.com>