lib: TCK: Override the default memory size for libguestfs appliance
Before the TCK test suite is even run, a dummy VM image is created with
virt-builder from a template. The problem is that the templates are
never updated since the moment they are published, so even a single
package install can trigger update to some packages. The problem is the
default memory size for the libguestfs appliance that can be queried
with "guestfish get-memsize".
On CentOS-8 this corresponds to 500M which may not be enough for DNF and
our test suite would keep failing on DNF timing out in the virt-builder
install step, see the BZ[1]. Since the install passes on Fedora-32 with
the default appliance memory of 768M and also since we're only
installing a single package (dsniff), the install will not trigger a
massive update, so using a 1G memsize override should suffice easily.

BZ tracking the default memsize in libguestfs:
https://bugzilla.redhat.com/show_bug.cgi?id=1837765

Signed-off-by: Erik Skultety <eskultet@redhat.com>