nwfilterxml2fwallout: Convert all matched text to lowercase
The unfortunate thing about nwfilter testing in TCK is that it's mostly
a shell machinery wrapped by the tck binary and as such it compares the
expected output to the verbatim output of tools like iptables. This is
very error prone for many reasons, but fixing it properly would require
a complete rework. Hence, at least make sure that we always only
compare lowercase text because the case can (and indeed did in the past)
change with iptables releases.
The culprit in question this time is the ipt-no-macspoof-test.

Signed-off-by: Erik Skultety <eskultet@redhat.com>