nwfilter: ip-spoofing: Change the IP address on an active interface
Modern distros like Fedora will have NetworkManager in place which will
track the state of the devices and hence apply some built-in
configuration when the device changes states. In context of this test
case it would automatically run DHCP renewal once we changed the IP
address on a disabled interface and brought it back up which in turn
would break the IP spoofing setup we wish to test with. Therefore,
flush and apply new IP configuration on an active interface.

Signed-off-by: Erik Skultety <eskultet@redhat.com>