ci: rationalize set of build platforms
Currently the CI jobs run across every build platform known to the
libvirt-ci project.

In terms of covering libvirt-tck functionality this is serious
overkill. We only ever test against libvirt / libvirt-perl git
master, so the most important thing to validate is the Perl
runtime version. For this it suffices to test with the oldest
expected version via CentOS 8, and the newest expected version
via Fedora Rawhide.

There are a few pieces related to RPMs in the Module::Build
recipe, so it is also worth testing a non-RPM platform via
Ubuntu 20.04.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>