ci: introduce use of lcitool manifest
Introduce a ci/manifest.yml file that defines the set of containers
and build jobs to be used for CI purposes. The dockerfiles get
renamed due to the different naming conversion used by the lcitool
'manifest' command that auto-generates all needed files.

The main .gitlab-ci.yml file now only needs to define the basic
build recipe which is project specific. The jobs are all generated
automatically from the manifest information.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>