DomainBuilder.pm: use "unsafe" cache for qemu disks
When testing QCOW2 images that need to be passed as RAW to QEMU,
the following error will be noticed IF the image is opened with
flag O_DIRECT (cache=none).

qemu-system-x86_64: -device virtio-blk-pci,bus=pci.0,addr=0x3,drive=libvirt-1-format,id=virtio-disk1,write-cache=on: Cannot get 'write' permission without 'resize': Image size is not a multiple of request alignment

That's because, from the RAW driver point of view, the actual
qcow2 size is unaligned (just metadata stored) and fails in the
condition introduced by QEMU's 9c60a5d1978.

It fixes the tests:
- 200-qcow2-single-backing-file.t
- 210-qcow2-auto-probing.t

Signed-off-by: Jose R. Ziviani <jziviani@suse.de>