scripts: Override the config path variable for networkApplyTest.sh
The libvirt-tck harness takes care of canonicalizing the config path,
but if someone were to use Avocado as the harness, it runs the Perl
scripts directly without taking care of canonicalizing the path, simply
because it doesn't have any knowledge about what types of values it
passes to the test scripts.
The networkApplyTest script expects the harness to change working
directory before executing the test cases which means that specifying
the config as a relative path won't work. At the same time though we
can't specify the config path as absolute in the avocado hints file
because we don't know the environment prior test execution.

Note that this problem only exists for tests which wrap the
networkApplyTest Bash script, because the native Perl test can
interact with the TCK.pm module to give them config values.

Signed-off-by: Erik Skultety <eskultet@redhat.com>