scripts: networkApplyTest: Simplify the 'sed'
The proposed replacement isn't exactly a oneliner, but it surely
doesn't force the reader to go and read 'info sed' in order to
understand the usefulness of the pattern & hold buffers and their
respective action commands.
Also, we don't target performance with this test suite, so the tiny
Bash penalty for chaining a few commands with a pipe isn't a problem.

Signed-off-by: Erik Skultety <eskultet@redhat.com>