nwfilter/220-no-ip-spoofing.t: Update for newer guest distro
Commit 7a332e940d6b bumped the guest image from Fedora 38 to 40 and
roughly since then our integration tests started failing.  The test
which check that denying IP spoofing works counts with the fact that IP
address for the interface is updated using dhclient, but that is no
longer the case.  So instead of killing dhclient use nmcli to set the IP
address and restore it properly.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>