TCK: default to "fullos" when kernels are not configured
Currently, behavior of generic_domain() and generic_machine_domain()
regarding the "fullos" argument is the following:

 - if "fullos" is explicitly specified, use its value
 - if not, set it to 0 by default, i.e. use direct kernel boot

This behavior is not optimal as not every driver supports direct kernel
boot. Additionally, for most of the tests using direct kernel boot or
fullos boot doesn't make any difference, esp. for cases when the
"fullos" argument is not passed.

To make that more flexible, default to "fullos = 0" only if there's at
least a single kernel configured in the configuration file, otherwise
use fullos mode.

Signed-off-by: Roman Bogorodskiy <bogorodskiy@gmail.com>