networks/400-guest-bandwidth.t: skip for bhyve
Bhyve doesn't support network bandwidth control, and the test itself is
bound to the tc(8) tool. Thus, there's a lot of work to make it pass, so
just skip it early for bhyve.

Also, updated number of skipped tests to 8 for no available IPv4
scenario as it probably should be the same as for the bhyve case.

Signed-off-by: Roman Bogorodskiy <bogorodskiy@gmail.com>