domain/120-disks-stats.t: wait for non-zero block stats
Currently, the code makes only a single attempt to fetch block
statistics before checking if it's non-zero. That might be not enough
for some drivers, e.g. for bhyve, as it needs some time to get
non-zero values.

Update the code to wait for non-zero block statistics for up to 60
seconds.

Signed-off-by: Roman Bogorodskiy <bogorodskiy@gmail.com>